WebSettings: Refactor rendering code
