+++ TF_Settings_Web/src/Components/Pages/Camera/Masking/MaskingScreen.tsx
@@ -14,8 +14,9 @@ import { Mask } from 'TouchFree/Tracking/TrackingTypes';
 import MaskingLensToggle from './MaskingLensToggle';
 import MaskingOption from './MaskingOptions';
 import { MaskingSliderDraggable, SliderDirection } from './MaskingSlider';
-import { setupRenderScene, updateCanvas } from './displayLensFeeds';
-import { HandState, rawHandToHandData, setHandRenderState } from './handRendering';
+import { updateCameraCanvas } from './createCameraData';
+import { rawHandToHandData, setHandRenderState } from './createHandData';
+import { setupRenderScene, updateHandRenders } from './sceneRendering';
 
 export type Lens = 'Left' | 'Right';
 
@@ -24,7 +25,6 @@ const FRAME_PROCESSING_TIMEOUT = 30;
 const MaskingScreen = () => {
     // ===== State =====
     const mainLens = useStatefulRef<Lens>('Left');
-    const handData = useStatefulRef<HandState>({});
     // Config options
     const masking = useStatefulRef<Mask>({ left: 0, right: 0, upper: 0, lower: 0 });
     const isCamReversed = useStatefulRef<boolean>(false);
@@ -64,6 +64,7 @@ const MaskingScreen = () => {
     const camFeedRef = useRef<HTMLDivElement>(null);
     const successfullySubscribed = useRef<boolean>(false);
     const frameTimeoutRef = useRef<number>();
+    const hasHandRenders = useRef<boolean>(false);
     const handTimeoutRef = useRef<number>();
 
     // ===== Hooks =====
@@ -149,7 +150,7 @@ const MaskingScreen = () => {
             successfullySubscribed.current = true;
 
             frameTimeoutRef.current = window.setTimeout(() => {
-                updateCanvas(data, mainLens.current, isCamReversed.current, showOverexposed.current, handData.current);
+                updateCameraCanvas(data, mainLens.current, isCamReversed.current, showOverexposed.current);
                 isFrameProcessing.current = false;
             }, FRAME_PROCESSING_TIMEOUT * 2);
         } else if (!successfullySubscribed.current) {
@@ -163,12 +164,15 @@ const MaskingScreen = () => {
         isHandProcessing.current = true;
 
         const hands = evt.detail?.Hands;
-        if (hands.length > 0 || handData.current.one || handData.current.two) {
+        if (hasHandRenders.current || hands.length > 0) {
             const handOne = hands[0];
             const handTwo = hands[1];
             const convertedHandOne = handOne ? rawHandToHandData(handOne) : undefined;
             const convertedHandTwo = handTwo ? rawHandToHandData(handTwo) : undefined;
-            handData.current = { one: convertedHandOne, two: convertedHandTwo };
+
+            hasHandRenders.current = convertedHandOne !== undefined || convertedHandTwo !== undefined;
+
+            updateHandRenders({ one: convertedHandOne, two: convertedHandTwo });
         }
 
         // Ignore any messages for a short period to allow clearing of message handling
